preparing to work with stable coin

+++ Data/DataTypes/PositionInfo.cs
@@ -4,7 +4,7 @@
     {
         public bool InPosition = false;
         public bool IsLong = false;
-        public int Quantity = 0;
+        public double Quantity = 0;
         public double Price = 0;
         public DateTime OpenDate = DateTime.Today;
         public string Symbol = "";
